feat(dc): add coords to photo [8140-35]

+++ lib/model/department_control/dcphoto.dart
@@ -11,8 +11,8 @@ abstract class DCPhoto with _$DCPhoto {
     String name,
     String data,
     String loadDate,
-    int geometryX,
-    int geometryY,
+    double geometryX,
+    double geometryY,
   }) = _DCPhoto;
 
   factory DCPhoto.fromJson(Map<String, dynamic> json) =>
